resultat4

+++ database/migrations/2024_06_28_144233_create_imports_table.php
@@ -14,7 +14,7 @@ public function up() {
     Schema::create('imports', function (Blueprint $table) {
         $table->id();
         $table->integer('quantity');
-        $table->decimal('value', 15, 2);
+        $table->integer('value');
         $table->string('country_origin');
         $table->date('departure_date');
         $table->date('arrival_date');
